requirements.txt
fastapi
uvicorn
pyJWT
firebase
firebase-admin
python-dotenv
google-cloud-firestore